requirement.txt
openai
streamlit
langchain
pyyaml
tiktoken
python-dotenv